refactor(services): complete migration to shared/ and remove all shims
- Update all imports to use shared/ paths directly
- Delete all re-export shim files
- Update _SYMBOL_MODULES mapping in services/__init__.py
- Update test mock paths to shared/ modules

No backward compatibility shims - all imports must use new paths

Closes #2393 #2394
Related to #2099
