--- src/vibe3/services/loc_service.py
@@ -1,9 +1,0 @@
-"""Re-export shim for LOC service.
-
-The actual implementation has moved to
-vibe3.services.shared.loc.
-"""
-
-from vibe3.services.shared.loc import LocService, LOCStats
-
-__all__ = ["LOCStats", "LocService"]
