+++ src/vibe3/services/pr/verdict_service.py
@@ -13,11 +13,11 @@ from loguru import logger
 
 from vibe3.clients import GitClient, SQLiteClient
 from vibe3.models import VerdictRecord, VerdictValue
-from vibe3.services.actor_support import extract_role_from_actor
 from vibe3.services.flow_service import FlowService
 from vibe3.services.handoff.storage import HandoffStorage
+from vibe3.services.shared.actors import extract_role_from_actor
 from vibe3.services.shared.paths import GitPathProtocol
-from vibe3.services.signature_service import SignatureService
+from vibe3.services.shared.signatures import SignatureService
 
 
 class VerdictService:
