+++ tests/vibe3/services/test_verdict_service.py
@@ -6,8 +6,8 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from vibe3.models.verdict import VerdictRecord
-from vibe3.services.actor_support import extract_role_from_actor
 from vibe3.services.pr.verdict_service import VerdictService
+from vibe3.services.shared.actors import extract_role_from_actor
 
 
 class TestVerdictService:
